Fix usages of FutureBuilder (#2128)
This revealed a bug involving screen_parent_resizer: The map always overflowed, without contracting itself again.
We introduced screen_parent_resizer due to a bug in maplibre, namely that half of the screen was unresponsive after an orientation change (see #415).
I could not reproduce this bug anymore, so it is likely safe to just remove the workaround altogether.